BugFix: allow font-load from sub-directory w/versions without bundled fonts
I, perhaps carelessly, disabled the code to install fonts from the place
where Mudlet extracts the bundled fonts to and then loads them in (in
versions WITH bundled fonts) to be done for the versions WITHOUT those
fonts.  This means that a module that includes a font for a particular MUD
could not rely on Mudlet making a font placed in that part of the mudlet
area of the user's file-system from loading the font in.

This commit remedies that error I hope.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>