Revise: change to NOT do scaling when drawing room symbols from cache
The previous QPainter::drawPixmap(...) performed a scaling operation to
make the symbol pixmap fit the specified rectangle. This is the cause
behind the poor rendering of text characters as the scaling undoes the
benefits of anti-aliasing and take time to do. This commit should be faster
because the pixmaps are generated at the size/resolution they are needed
(though they do have to be thrown away and regenerated if the zoom or other
sizing factors change) - they do however look better to me!

Also:
* merged (int) TRoom::xzoom and TRoom::yzoom into TRoom::xyzoom.
* added the symbol scaling "fudge-factor" to the "Special Options" tab of
  the "Profile preferences" dialog - it may be helpful to artificially
  over-size (> 1.00) or under-size (< 1.00) the symbols in some situations.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>