BugFix: fixup Host class initialisation (fixes #1561)
Reorder Host constructor initialisation list to be correct compared to the
header file declaration of members.

Added initialisers for and with the value shown (and thus solve an issue
which exists because Host::mIsNextLogFileInHtmlFormat was not initialised):
* (bool) Host::mIsNextLogFileInHtmlFormat(false)
* (bool) Host::mShowInfo(true)
* (bool) Host::mIsProfileLoadingSequence(false)
* (bool) Host::mShowPanel(true)
* (bool) Host::mIsClosingDown(false)

removed unused:
* (int) Host::mMainIconSize
* (QString) Host::mMudOutputBuffer
* (bool) Host::mShowToolbar
* (int) Host::mTEFolderIconSize
* (QStringList) Host::mTextBufferList - this was probably intended to store
  the hidden TConsole instances that are only used to hold text that is
  never shown.
* (QStringList) mParagraphList

Also take out a couple of null class Exception definitions in the Host
class header which do some to be completely disassociated from anything...?

Leaving in place although not currently used:
* (int) Host::retries
* (int) Host::timeout
* (QString) Host::userDefinedName
the first pair were probably intended to handle connection issues but no
connection time out / retry code was developed (thought I can see potential
for them - e.g. for countering non-user initiated Mud Server disconnects).
The last would allow a user to give a profile a nick-name and allow the
profile to be renamed WITHOUT having to modify the underlying file-system.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>