Revise: prevent qmake spamming Windows builds unnecessarily
Some build configuration details only need to be reported on the first
pass of the three that occur on Windows (or any other platform where
the qt CONFIG variable contains "debug_and_release").  On the first pass
"build_pass" is NOT present in the CONFIG but it and "debug" or "release"
as appropriate is added to that variable.

The trouble is that I made a previous typo and used "build-pass" instead,
this commit corrects that.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>