BugFix: correct fatal issue when 4-byte GB18030 sequences detected in GBK
Due to failing to set a flag (isToUseReplacementMark) in a section of
(bool) TBuffer::processGBSequence(...) the QChar/TChar counts for any line
containing 4-byte GB18030 encoding characters (which need to be treated as
invalid if the mode is set to only 2-byte GBK) would be messed up by
missing replacement character marks which quickly causes a crash when the
display code runs out of QChars whilst painting the screen.

I also extended the code controlled by the DEBUG_GB_PROCESSING #define so
that it actually reported which range of characters the decoded GBK
equivalent to a Unicode "Codepoint" falls within.

I have also undefined DEBUG_GB_PROCESSING and DEBUG_UTF8_PROCESSING so they
will have to be redefined to show some of the above changes, but will not
make the stdout/stderr output so spammy if the UTF-8, GBK or GB18030
encodings are selected.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>