Enhance: add text analyser to consoles
This will allow the user to determine the exact characters present in a
line of MUD (or locally generated) text which could be useful in working
out what to use to construct triggers and also how the displayed characters
are constructed in UTF-8 bytes.  To use it the user should control click
on a line of text in a console and then right click to bring up the context
menu.  By moving the cursor over the "analyser characters" action and
hovering there a generated tool-tip will be brought up that contains HTML
tables with the break down of Unicode code-points which is then broken down
into the corresponding UTF-16 (which Qt uses internally to store
all text in QStrings) and UTF-8 which is what the Lua sub-system gets as
its form of strings.  For characters that are not on the Basic Multi-Plane
and which thus need two QChar (and thus UTF-16 surrogate pairs) to
represent the pairs of UTF-16 values (in the range 0xD800 to 0xDBFF for the
first, high surrogate followed by one in the range 0xDC00 to 0xDFFF for the
second, low surrogate) those numbers are also displayed.

Note: this works on (only) the whole first line of the selected text even
if more than one line is selected - or only a fraction of that first line
is selected.  The context menu entry is NOT shown if there is no selection.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>