Use out-ouf-source builds to avoid issues with moc
This is (hopefully) more stable than deleting directories. Recent
changes also made it possible to get rid of copying dependencies, which
gets rid of another error prone process. The only shaky thing (to my
eyes) is, that Mudlet has to be run from the build folder (not
build/debug !!) to find the LuaGlobal entrypoint.