Enhance: add 16M color mode support
This commit adds support for the limited reinterpretation of the ANSI
specification of the SGR [3|4]8;2;rrr;ggg;bbb ... m colour code that is now
used to specify one of 16M colours directly.  It does not support the
proper form that is supposed to handle up to 7 sub-options after the
initial 38 or 48 where the sub-options are separated with the otherwise
reserved ':' character - and which if the second argument is 2 or 3 or 4
can have a VARIABLE number of arguments with the omitted ones to be taken
as zero.  The absence of the full, proper, form means that the next three
numbers MUST be read as the red, green and blue numbers - although the
implementation here can fill in missing ones as zeros if the terminating
'm' character is seen rather than the expected numbers.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>