Enhance: indicate other profile with new data when multi-playing
If more than one profile is open and multiView mode is not selected then
any profile which is not the currently selected one will be hidden from
view - and if any data comes in from the MUD server it will not be obvious.

This commit turns the text on the tab of the "other" profile red when new
data arrives - which is reset when the user switches to that profile to see
what has come in.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>