revise: use icons as well as colour to show activity on other Tabs
This attempts to detect some local initiated things that change the
contents as well as incoming data from the Mud Server, in the former case
a blue-which info icon is added to the tab for the affected, "other"
non-active profile, which also turns the title text blue to match.  For
data coming in from the Mud Server an orange warning triangle is placed
there together with the red text.  This has priority over the local case.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>