Revise: add ability to set individual Tabs to have bold text on QTabBar
Some code was suggested and modified to sub-class the QTabBar class into
one (TTabBar) with the ability to set individual tabs to have bold text on
a per tab (selected by text OR index). This also required the derivation of
a TStyle from QProxyStyle to perform the actual drawing of tabs so
modified.

I have removed the previous use of icons and changed the red on data from
the MUD server to magenta.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>