Revise: use GITHUB instead of SOURCEFORGE as source of zziplib
SF is becoming increasingly unreliable as a host for source code and this
project has moved to Github some time ago but we had not spotted this.

There are later versions that the 2012 0.13.62 version that we are
currently using but that is an exercise for a new PR instead of a hot-fix
repair to the system for this one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>