Simplify: only respond to remote server incoming data and only bold tab
This actually makes the TConsole code a tiny bit more complex as it has
to filter whether to emit a SIGNAL depending on where the input to
TConsole::printOnDisplay(...) comes from!

I was also able to simplify the mudlet::slot_newDataOnHost(...) method
because I had forgotten to use the methods to change the tab text
formatting that identifies the tab by its text as an argument rather than
an index number. (The same for(...) loop construct is used but it is hidden
in the new TTabBar class...)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>