Enhance: indicate other profile with new data when multi-playing (#1589)
This commit is a squash down one, an edited summary of the component
commits that made it up is:

If more than one profile is open and multiView mode is not selected then
any profile which is not the currently selected one will be hidden from
view - and if any data comes in from the MUD server it will not be obvious.

This commit modifies the text on the tab of the "other" profile when new
data arrives - which is reset when the user switches to that profile to see
what has come in.

A revision attempts to detect some local initiated things that change the
contents as well as incoming data from the Mud Server.

Some code was suggested and modified to sub-class the QTabBar class into
one (TTabBar) with the ability to set individual tabs to have bold text on
a per tab (selected by text OR index). This also required the derivation of
a TStyle from QProxyStyle to perform the actual drawing of tabs so
modified.

I have revised the new code to support italic and underline effects as well
as bold, although these are currently utilised.

I was also able to simplify the mudlet::slot_newDataOnHost(...) method
because I had forgotten to use the methods to change the tab text
formatting that identifies the tab by its text as an argument rather than
an index number. (The same for(...) loop construct is used but it is hidden
in the new TTabBar class...)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>