Change source for zziplib sources (#1592)
* Add zziplib 0.13.62 sources to 3rdparty dir

The sourceforge downloads have bcame too unreliable to use and the
new Github sources don't provide certain needed files. This is a
temporary solution until we can update to a later version that gets
rid of dependencies to create the needed files which we can not
satisfy on windows. See
https://github.com/gdraheim/zziplib/issues/8

* iAdapt install script to included zziplib sources