Show changelog on startup with the update news
First go at showing the changelog on startup with some content instead of a blank one. In order to do this, we need to record the previous version we upgraded from (in case of skipped versions) and then tell that to the changelog dialog so it can generate the appropriate text.