Revise: further simplifications + dead code removal
Some TTextEdit methods were using indirections through pointers that
actually resolve to the 'this' one so could be removed.

The TTextEdit::isTailMode() returns a value that is already public, it does
not do anything that is otherwise effective and it is only used once
outside of the class so it is redundant and can be remove IMO.

In the TTextEdit::scroll{Up|Down}(int) methods there was dead code that
modified the local copy that is made of the supplied argument but no
effective use was made of that local copy. Also the scrollUp method was
setting the TTextEdit::mIsTailMode flag to true for the lower pane - but
that is always true and never modified anyhow.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>