Enhance: sort out non ANSI code numbering in saved files and editor
Fix the mis-handling of colour trigger numbers (because the previously used
numbers did not match the "ANSI" colour codes) in the first commit; as part
of this I have replace the pattern (matched with the regexp
"FG(\d+)BG(\d+)" used to encode the ANSI colours used for the textual
representation of a colour pattern item in a TTrigger with:
"^ANSI_COLORS_F{(\d+|DEFAULT|IGNORE)}_B{(\d+|DEFAULT|IGNORE)}$"

The processing of this pattern is abstracted to:
* (QString) TTrigger::createColorPatternText(const int, const int)
* (void) TTrigger::decodeColorPatternText(const QString&, int&, int&)

Extends the colour trigger dialog to better handle the existing 16 ANSI
colours but now also handles the remainder of the 256 ANSI colour code
numbers and provides a means to use the "default" fore or background colour
(what ever the user has set it to) and also to ignore either the fore or
background colours.

Refactored the mapping of the colour buttons in the dlgColorTrigger
form/dialog to use a QSignalMapper which removed the need for 16 almost
identical methods (one for each button) and added text to them to describe
the colours they are supposed to represent even if they have been
completely remapped to other colours by the user - or, in the future if
we include support for OSC (Operating System Command) Server setting of
colour by use of \033]4;<colour number>;<colour specification>; ... ;\007
to set the colour with the given number to the X11 name, or a 6 hex digit
RRGGBB string; with \003]104;<colour number>;...;\007 to reset them to
"standard". These come from the XTerm terminal emulator:
http://invisible-island.net/xterm/ctlseqs/ctlseqs.html#h2-Operating-System-Commands
Also added a couple of helpers to set the background colour on a QButton to
the given colour and the foreground to a contrasting colour of either
black or white given the QColor::lightness() of that colour:
* (static QString) dlgTriggerEditor::generateButtonStyleSheet(const QColor&)
* (static QColor) dlgTriggerEditor::parseButtonStyleSheetColors(
                      const QString&, const bool isToGetForeground = false)

A further enhancement means that editing the above colour pattern string
when the trigger item type is set to NOT be a colour trigger is picked up
and converted to the correct colour in the dialogue if the colour trigger
type is re-selected later.

To support the full range of 256 ANSI colours without breaking existing
scripts it is necessary to provide a new:
(int) TLuaInterpreter::tempAnsiColorTrigger(lua_State*)
that uses a different mapping of numbers in the range 0 to 15 and adds
-1 for ignore one of either the fore or background colours (both is a run-
time error!)

Add a (QColor) Host::getAnsiColor(const int, const bool isBackground) const
to return the correct QColor value from the Host instance given an ANSI
(0 to 255) colour code number but also supports -1 for the default fore or
background colour for that Host instance.

Also:
* rename (void) TTrigger::set[BF]gColor(QColor&) to:
  (void) TTrigger::setColorizer[BF]gColorsetFgColor(QColor&) because it
  helps to link them to the colouring of matching text and separates them
  from methods relating to colour pattern matching.

* Remove a memory leak where TColorTable structs were not being deleted
  when they were being removed from the:
  (QList<TColorTable*>) TTrigger::mColorPatternList
  in the TTrigger destructor and when it was being updated in
  (bool) TTrigger::setRegexCodeList(QStringList, QList<int>)

Remove:
* (bool) TTrigger::mColorTrigger[BF]g as they are not needed and can be
  derived from the related (int) TTrigger::mColorTrigger[BF]gAnsi being
  not equal to the new TTrigger::scmIgnored constant value.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>