BugFix: correct some errors Codecy pointed out & optimise TBuffer for-loops
I was using multiple post-increment operators on the same variable in some
method calls but did not realise that this was undefined behaviour because
the order of processing is not specified in TBuffer.cpp. I also made a
similar error in (void) mudlet::slot_updateBlinkPhase().

Whilst searching for uses of the operator++ in TBuffer I found several
cases, including in some for-loops where the postfix form could be replaced
by the more efficient prefix form - and/or merged into another operation.

There was some unreachable code at the start of:
(QString) TBuffer::bufferToHtml(const bool showTimeStamp, const int row,
             const int endColumn, const int startColumn,  int spacePadding)
which has been removed.

Added run-time error message to Lua commands that set display attributes
so that trying to apply an attribute on a non-existent window is reported.

Fixed a couple of "exceptional" spelling mistakes!

Also removed three unused members that Codecy detected as being
uninitialised:
* (int) TBuffer::speedTP
* (int) TBuffer::speedSequencer;
* (int) TBuffer::speedAppend;

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>