Revise: polish up some aspects of AlysCole's PR
Change the strings use to format the log file name to remove the risky
':' character - this is used in special ways on Windows OS (Drive
selection, NTFS alternate data streams).

Consider reading, 8-) :
https://www.dwheeler.com/essays/fixing-unix-linux-filenames.html

Makes the Host::mLogFileNameFormat being empty as the signal value to use
the Host::mLogFileName itself as the actual logfile name and make the
latter being empty use the "placeholder" text. Combined this eliminates
the need to store the index into the
dlgProfilePreferences::comboBox_logFileNameFormat control in the profile
data.

Make the sample log file names in the
dlgProfilePreferences::comboBox_logFileNameFormat control include an
extension and make that reflect the current format choice elsewhere on the
preference tabWidget.  This used a new
(void) slot_changeLogFileAsHtml(const bool isHtml) . To allow the user
specified filename to also show that extension add a QLabel to show it and
realign the controls and labels on that tabWidget so that the filename and
the extension abut each other.

Move the setup in dlgProfilePreferences of:
dlgProfilePreferences::comboBox_logFileNameFormat
to the place where other Profile related items are - so that it appears
blank if that tab is accessed when no profile is loaded.  In a related
manner change enabling/disabling of the logging preferences controls to
be done by the groupBoxes containing them.

Add a clear division (insert "</div><div>" tags) between multiple appended
logs - which make it clear that there is a gap in the log data.

Add a full start and end date and time text to each log and arrange the
text creation to allow it to be translated as appropriate.

Ensure that the HTML text log file is written in the UTF-8 encoding it
is set to declare it is using by adding a QTextDecodec to the QTextStream,
otherwise a "System" locale is used and, at least on Windows, that could be
something else entirely!

Also:
* Remove the 'QIODevice::WriteOnly' option from the opening of the log file
  it implies 'QIODevice::Truncate' and is unneeded with 'QIODevice::Append'
* fixup some grid layout items in the ui/profile_preferences.ui file to be
  in order - which keeps the git changes smaller and slightly easier to
  track.
* rejig some translatable texts originally in the PR (and some nearby that
  I had forgotten about), to remove most HTML tags from the translations -
  this makes things nicer for subsequent translation work. NB. It leaves
  paragraph and text formatting codes in the translations because those can
  be affected/need changes for translations {it is bad practice to bold or
  italicise ideographic texts - they have different means to apply
  emphasis!}
* Add some missing if() code block '{'...'}'s.
* Switch to using the Qt directory selection dialog when selecting where
  to put log-files - this enables the user to see the files that they may
  want to append to - especially on Windows where the native directory
  selection dialog (brokenly ?) HIDES the files...!
* Refactor (void) TBuffer::log(int, int) to build up a QStringList of
  each line of log text rather than continually appending each new line to
  the previous one - this should be more efficient.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>