BugFix: allow C key to be used as a key-binding.
Following a short discussion on Discord Sarren Nardieu#9574 pointed out to
me (and Eraene#7438 confirmed before I did as well) that 'C' could not be
used in a key-binding.  I found that Qt::Key_C was being processed
specially so that the `<Ctrl>-C` combination could be handled but if the
`<Ctrl>` was not present in the `QKeyEvent` the `C` keypress was being
discarded.

This commit fixes things so that the exact combination is processed IF
there is a selection in the main `TConsole` instance but otherwise the
combination will be processed as "other" keys are.

NOTE: This showed that there are other keys that cannot be bound by
"Keys" items in Mudlet because they are also processed specially.  I have
modified ALL of them to exactly match on the combinations of modifiers
previously coded but if there is not a match then the combination will now
be processed "normally" - this opens up the other combinations of modifiers
for use in Key-bindings - although there are some that are intercepted by
Qt directly because they are bound to the main menu actions - these tend to
be `Alt` key combinations and include C (for the "Connection Preferences"
Dialogue") and P (for the "Profile Preferences" one).

I also tidied up some pointlessly repeated simple assignments in the same
event handler code and converted some places where `QString`s were being
set to empty raw string literals to use the `QString::clear()` method
instead.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>