Revise: modify previous commit to process key-bindings better
I erroneously made the previous commit run all the code for "normal" keys
on the specially handled ones in (bool) TCommandLine::event(QEvent*) when
I should have limited it to just that part associated with checking for
and then processing key-bindings found that match against the input.

I also had to add the Escape key to the arrow keys that are handled
specially by an event filter in the dlgTriggerEditor active when the "grab-
key" process is running because it, like the arrow keys, is normally used
by the Qt libraries and would not get to the event code that deduces what
key combination the user is entering for a key-binding.  The Escape key is
actually used to abort the key capturing process but it was not working to
do that in the past.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>