Revise: update Key code to be more like other items
This commit makes the enableKey(<name>) and disable(<name.) lua functions
return true only if a key with the given name was found (instead of when
any Key was present in the profile) - which is more like the handling for
other items such as Triggers and Aliases.

It also arranges for all temporary keys to be "blown-away" when the profile
is reset - which was not happening even though it does happen for other
item types.

Furthermore when the "Emergency Stop" is activated it now also disables all
the keys like it disables all timers/aliases/triggers.  I think this is
desirable because if the user sets a binding on a normal key with no
modifiers that it is impossible to type the character on that key into the
profile's command line - so being able to disable the interception of the
key by a temporary key (that will not show up in the editor) is just one
reason to add keys to what the control does...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>