BugFix: add option to control selection between wc_width()/wc_width_cjk()
This is a quick-fix to accommodate the fact that some locales (not
currently fully identified) require graphemes with the informative
"East Asian Width" Unicode property of type "Ambiguous" to be drawn as
narrow characters (the default case) and others (probably East Asian) as
wide ones.

The option is controlled (on a per profile basis) by a new checkbox on the
Profile preference's "Main Display" tab - it is stored within the profile
data so will persist between sessions.

NOTE: This commit finally puts the Q_OBJECT macro into play in the Host
class - so it is important to ensure the qmake part of the build process
is re-run after a build of a different branch of the code so that the moc
correctly records that that class is one that it has to process.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>