Credit: apply missing credits to creator of original contribution to Mudlet
Something went wrong in trying to apply a Pull-Request from myself onto a
PR Huadong Qi wanted to made to the Mudlet codebase which made it
un-applyible.  To solve this I have taken their original code, appended my
extra code and intend to apply the combination as a second PR which should
merge into the main repository in a straightforward manner.

I have also, in this commit, filled out the copyright lines on the files
that Huadong Qi modified and added the GPL template with their name onto
the two new files that has been inserted into the code base.  The main one
wcwidth.cpp is a modified copy of Markus Knuth (C) 2007 wcwidth.c which
bears the following licencing terms:
"Permission to use, copy, modify, and distribute this software for any
purpose and without fee is hereby granted. The author disclaims all
warranties with regard to this software.

Latest version: http://www.cl.cam.ac.uk/~mgk25/ucs/wcwidth.c"

As such this means that we can (need?) to re-license it under a GPL but it
is my considered opinion that this is approprate for our usage.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>