BugFix: solve recently introduce precedential errors
There is one line in TCommandLine::event(QEvent*) where I had thought that
'&' had a higher precedence than (was evaluated before) '==' but Coverity
pointed out that it was not.

At the same time I spotted a common copy/paste resulting in an extra space
in a consistent place throughout a section of the same code.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>