Testing: this version will suicide on some unlikely Room Exit dialog states
This version will produce a qFatal() message and die should the values
returned by examining the QButtonGroup for each door type control be not an
expected one.

I also spotted some less then perfect code which was triggering 2 Coverity
Medium level warning:
* 1414989 Explicit null dereferenced
* 1415026 Explicit null dereferenced

At the same time I saw that the dlgRoomExit::init() method is not needed
as it can be contained entirely within the constructor, and that we did
have a resource leak because the dialogue was not necessarily being removed
from memory when it was closed.

THIS IS A TEMPORARY COMMIT CONTAINING qFatal() TO CAUSE IT TO DIE SUDDENLY
THIS IS REFLECTED IN A BLOODY OBVIOUS SPLASH SCREEN CHANGE AND A [WARN]
MESSAGE A FEW SECONDS AFTER ANY PROFILE IS OPENED. IT SHOULD ONLY BE USED
OFF-LINE WITH AN EXISTING MAP TO TEST THE ROOM EXIT DIALOGUE.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>