Cleanup: remove debug code, but leave other tweaks inplace
The prior commit DID fix the original problem!

I was unclear on the precise nature of what the "default-constructed" value
would be for a QMap<int, T*>::value(int) call should the key not be found
in the map.  As my suspicions suggested a nullptr would be returned which
would not be helpful I now ensure an "empty" TExit instance is kept around
so that a pointer to it can be returned if necessary. This should prevent
any possibility of trying to deference a null-pointer by accident
afterwards.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>