Redo fix cjk fullwidth character display (#1668)
This is a squashed commit containing a number of changes:

Something went wrong in trying to apply a Pull-Request from myself onto a
PR Huadong Qi wanted to made to the Mudlet codebase which made it
un-applyible.  To solve this I have taken their original code, appended my
extra code and which is merge into the main repository in a straightforward
manner here.

I have also, in this commit, filled out the copyright lines on the files
that Huadong Qi modified and added the GPL template with their name onto
the two new files that has been inserted into the code base.  The main one
wcwidth.cpp is a modified copy of Markus Knuth (C) 2007 wcwidth.c which
bears the following licencing terms:
"Permission to use, copy, modify, and distribute this software for any
purpose and without fee is hereby granted. The author disclaims all
warranties with regard to this software.

Latest version: http://www.cl.cam.ac.uk/~mgk25/ucs/wcwidth.c"

As such this means that we can (need?) to re-license it under a GPL but it
is my considered opinion that this is appropriate for our usage.

Other notes from original commits:

TTextEdit::drawLine: add non-BMP character support.

Replace GetNextGraphemeSize with QTextBoundaryFinder

Add option to control selection between wc_width()/wc_width_cjk()

This is a quick-fix to accommodate the fact that some locales (not
currently fully identified) require graphemes with the informative
"East Asian Width" Unicode property of type "Ambiguous" to be drawn as
narrow characters (the default case) and others (probably East Asian) as
wide ones.

The option is controlled (on a per profile basis) by a new checkbox on the
Profile preference's "Main Display" tab - it is stored within the profile
data so will persist between sessions.

Leaving the profile preference control, now labelled:
"Make 'Ambiguous' E. Asian width characters wide"
in the tri-stated (partially-checked) default setting means that GBK and
GB18030 MUD server encodings will use the 'wide' setting and all others the
'narrow' one, but allows the user to manually force the setting to be
'narrow' (unchecked) or 'wide' (checked) if required.

NOTE: This commit finally puts the Q_OBJECT macro into play in the Host
class - so it is important to ensure the qmake part of the build process
is re-run after a build of a different branch of the code so that the moc
correctly records that that class is one that it has to process.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>