Update Edbee with my latest commit, comments from commit on edbee-lib below:
Updates the key handler to allow enter/return/tab to go to editor as usual if the currentWord exactly matches the selected item on the list(which it always will immediately after typing an exact match, based on the way the list gets updated as you type).
The menu also now hides if there is only one match in the list, and it's an exactly match of the currentWord.